exported resonator link to generic component
